Comment controller has been updated with extra validation

+++ server/routers/commentRoutes.js/comment.controller.js
@@ -5,23 +5,37 @@ const Recipe = require('../../models/recipes.mongo');
 //Adds a comment under a post - should be good
 async function httpAddComment(req, res){
     try{
-        let authorId =  await User.findOne({authId : req.body.author});
-        let newComment = await Comment.create({...req.body, author: authorId});
-        await User.findOneAndUpdate({
+        const authId = req.body.author;
+        if(!authId){
+            return res.status(404).json({msg: "An author is required"});
+        }
+        const authorId =  await User.findOne({authId : authId});
+        if(!authorId){
+            return res.status(404).json({msg: "Could not find the author"})
+        }
+        const newComment = await Comment.create({...req.body, author: authorId});
+        if(!newComment){
+            return res.status(404).json({msg: "Could nto create the comment"})
+        }
+        const updatedUser = await User.findOneAndUpdate({
             _id: authorId.id}, {
               $addToSet: {comments: newComment.id}
             }
         );
-        
-        await Recipe.findOneAndUpdate({
+        if(!updatedUser){
+            return res.status(404).json({msg: "User has not been updated"})
+        }
+        const updatedRecipe = await Recipe.findOneAndUpdate({
             _id: newComment.recipe
         }, {
             $addToSet: {comments: newComment.id}
         })
+        if(!updatedRecipe){
+            return res.status(404).json({msg: "Recipe has not been updated"})
+        }
         return res.status(201).json(newComment);
     } catch(err){
-        console.log('Errr' ,err)
-        return res.status(404).json({msg: "Unable to add a comment"})
+        return res.status(500).json({msg: "An error has occured, unable to add a comment"})
     }
 }
 
@@ -29,9 +43,21 @@ async function httpAddComment(req, res){
 async function httpDeleteComment(req, res){
     try{
         const id = req.params.id; 
-        let deletedComment = await Comment.findOneAndDelete({_id: id});
-        await User.findOneAndUpdate({_id: deletedComment.author}, {$pull: {comments: deletedComment._id}});
-        await Recipe.findOneAndUpdate({_id: deletedComment.recipe}, {$pull: {comments: deletedComment._id}})
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ msg: 'Invalid id' });
+        }
+        const deletedComment = await Comment.findOneAndDelete({_id: id});
+        if(!deletedComment){
+            return res.status(404).json({msg: "Comment could not be deleted"})
+        }
+        const userUpdated = await User.findOneAndUpdate({_id: deletedComment.author}, {$pull: {comments: deletedComment._id}});
+        if(!userUpdated){
+            return res.status(404).json({msg: "User has not been updated"})
+        }
+        const recipeUpadted = await Recipe.findOneAndUpdate({_id: deletedComment.recipe}, {$pull: {comments: deletedComment._id}})
+        if(!recipeUpadted){
+            return res.status(404).json({msg: "Recipe has not been updated"})
+        }
         return res.status(200).json(deletedComment);
     } catch(err){
         return res.status(404).json({msg: "Unable to delete a comment"})
@@ -41,12 +67,21 @@ async function httpDeleteComment(req, res){
 //Edits the comment for a user - Finished
 async function httpEditComment(req, res){
     try{
-        let commentId = req.params.id
-        let body = req.body
-        await Comment.findOneAndUpdate({_id: commentId}, body, {upsert: true} )
-        return res.status(201).json({msg: 'Edited the recipe'})
+        const commentId = req.params.id;
+        if(!commentId){
+            return res.status(404).json({msg: "Comments id is needed"})
+        }
+        const body = req.body;
+        if(!body){
+            return res.status(404).json({msg: "Body of a comment is required"})
+        }
+        const updatedComment =  await Comment.findOneAndUpdate({_id: commentId}, body, {upsert: true} )
+        if(!updatedComment){
+            return res.status(404).json({msg: "Comment has not been updated"})
+        }
+        return res.status(201).json({msg: 'Comment has been updated'})
     } catch(err){
-        return res.status(404).json({msg: "Unable to edit a comment"})
+        return res.status(500).json({msg: "An error has occured, unable to edit a comment"})
     }
 }
